longer update

+++ GitHW1/src/Main.java
@@ -1,9 +1,9 @@
 public class Main {
     public static void main(String[] args) {
 
-        SalesManager salesManager = new SalesManager(new int[]{1,29,4,12,13,19,21,25,26,27});
+        SalesManager salesManager = new SalesManager(new long[]{1,29,4,12,13,19,21,25,26,27});
 
-        int maxElement = salesManager.max();
+        long maxElement = salesManager.max();
 
         System.out.println(maxElement);
     }
